# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GuestRsvpStatus = typeof GuestRsvpStatus[keyof typeof GuestRsvpStatus];


export const GuestRsvpStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  declined: 'declined',
} as const;

export interface Guest {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  tableNumber?: number | null;
  plusOne?: boolean;
  /** @nullable */
  plusOneName?: string | null;
  invitationToken: string;
  rsvpStatus: GuestRsvpStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface GuestInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  tableNumber?: number;
  plusOne?: boolean;
  plusOneName?: string;
  notes?: string;
}

export type GuestUpdateRsvpStatus = typeof GuestUpdateRsvpStatus[keyof typeof GuestUpdateRsvpStatus];


export const GuestUpdateRsvpStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  declined: 'declined',
} as const;

export interface GuestUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  tableNumber?: number;
  plusOne?: boolean;
  plusOneName?: string;
  rsvpStatus?: GuestUpdateRsvpStatus;
  notes?: string;
}

export interface WeddingConfig {
  id: number;
  brideName: string;
  groomName: string;
  weddingDate: string;
  venue: string;
  venueAddress: string;
  dressCode: string;
  allowedColors: string[];
  ceremonyTime: string;
  receptionTime: string;
  /** @nullable */
  additionalInfo?: string | null;
  updatedAt?: string;
}

export interface Invitation {
  guest: Guest;
  weddingConfig: WeddingConfig;
}

export type RsvpUpdateRsvpStatus = typeof RsvpUpdateRsvpStatus[keyof typeof RsvpUpdateRsvpStatus];


export const RsvpUpdateRsvpStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  declined: 'declined',
} as const;

export interface RsvpUpdate {
  rsvpStatus: RsvpUpdateRsvpStatus;
  plusOneName?: string;
}

export interface WeddingConfigUpdate {
  brideName?: string;
  groomName?: string;
  weddingDate?: string;
  venue?: string;
  venueAddress?: string;
  dressCode?: string;
  allowedColors?: string[];
  ceremonyTime?: string;
  receptionTime?: string;
  additionalInfo?: string;
}

export interface Stats {
  totalGuests: number;
  confirmed: number;
  declined: number;
  pending: number;
  withPlusOne: number;
  totalAttendees: number;
}