# lib/api-zod/src/generated/types/guest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */
import type { GuestRsvpStatus } from './guestRsvpStatus';

export interface Guest {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  tableNumber?: number | null;
  plusOne?: boolean;
  /** @nullable */
  plusOneName?: string | null;
  invitationToken: string;
  rsvpStatus: GuestRsvpStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}