# lib/api-zod/src/generated/types/guestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */

export interface GuestInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  tableNumber?: number;
  plusOne?: boolean;
  plusOneName?: string;
  notes?: string;
}