# lib/api-zod/src/generated/types/invitation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */
import type { Guest } from './guest';
import type { WeddingConfig } from './weddingConfig';

export interface Invitation {
  guest: Guest;
  weddingConfig: WeddingConfig;
}