# lib/api-zod/src/generated/types/rsvpUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */
import type { RsvpUpdateRsvpStatus } from './rsvpUpdateRsvpStatus';

export interface RsvpUpdate {
  rsvpStatus: RsvpUpdateRsvpStatus;
  plusOneName?: string;
}