# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */

export interface Stats {
  totalGuests: number;
  confirmed: number;
  declined: number;
  pending: number;
  withPlusOne: number;
  totalAttendees: number;
}