# lib/api-zod/src/generated/types/weddingConfigUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wedding invitation management API for Katheryne & Rainer
 * OpenAPI spec version: 0.1.0
 */

export interface WeddingConfigUpdate {
  brideName?: string;
  groomName?: string;
  weddingDate?: string;
  venue?: string;
  venueAddress?: string;
  dressCode?: string;
  allowedColors?: string[];
  ceremonyTime?: string;
  receptionTime?: string;
  additionalInfo?: string;
}